+++ src/Service/BaseDelete.php
@@ -3,7 +3,7 @@
   * BaseDelete.php
   * Description: Principal object delete class of all templates
   * @Author : M.V.M.
-  * @Version 1.0.12
+  * @Version 1.0.14
 **/
 
 declare(strict_types=1);
@@ -14,6 +14,9 @@ use Psr\Log\LoggerInterface;
 use App\Service\BaseRepository;
 use App\Controller\BaseParameters;
 use PDOException;
+use PHPMailer\PHPMailer\PHPMailer;
+use PHPMailer\PHPMailer\SMTP;
+
 /*
    Observations :
    ROUTE : $group->delete('/{$tableName}/delete/{id}','App\Controller\{$tableName}\Delete:delete');
@@ -23,7 +26,6 @@ final class BaseDelete extends BaseRepository
 {
    protected object $tableClass;
    protected BaseParameters $parameters;
-   protected string $userInfo = '';
 
    public function __construct(object $tableClass, BaseParameters $parameters)
   {
@@ -41,13 +43,15 @@ final class BaseDelete extends BaseRepository
          $params  = [$this->tableClass->getFieldsId() => $id];                                             // Pasamos el id a parametros
          $result  = (array) $this->count( $this->parameters->getDb(),$this->tableClass->toTable(), $this->tableClass->getFieldsId(),$params);  //  leemos DDBB con id
          $msgDebug = 'count: '.$result['count'].' key: '.json_encode($params,JSON_PARTIAL_OUTPUT_ON_ERROR );
-         $this->toDebugger($userInfo, $this->parameters->getLogger(), $msgDebug, $this->query);                 // Si debug = true, graba el sql y parametros en el logger
-         if ($result['code'] == 200 &&  $result['count'] > 0)                                        // Si existe el registro con el select count(*)
+         $this->toDebugger($userInfo, $this->parameters->getLogger(), $msgDebug, $this->query);            // Si debug = true, graba el sql y parametros en el logger  
+         if ($result['code'] == 200 &&  $result['count'] > 0)                                              // Si existe el registro con el select count(*)
          {
             $sql  = "DELETE FROM `{$this->tableClass->toTable()}` WHERE `{$this->tableClass->getFieldsId()}`" . " = :" . $this->tableClass->getFieldsId(); // Delete DDBB
             $stmt = $this->parameters->getDb()->prepare($sql);                                                            // preparamos la sentencia sql
             $msgDebug = 'sql: '.$sql.' params: '.str_replace('"','',json_encode($params,JSON_PARTIAL_OUTPUT_ON_ERROR ));
             $this->toDebugger($userInfo, $this->parameters->getLogger(),'',$msgDebug);                                              // Si debug = true, graba el sql y parametros en el logger
+            $msgMail = $this->toMailer($this->parameters->getMailer(), 'delete' , $msgDebug);
+            if (!empty($msgMail)) { $this->toDebugger($userInfo, $this->parameters->getLogger(), 'Mailer message: ', $msgMail); }
             try
             {
                $stmt->execute($params);                                   // Ejecutamos la sentencia
@@ -88,4 +92,23 @@ final class BaseDelete extends BaseRepository
             $logger->debug($msg);
       }
     }
+
+    private function toMailer(PHPMailer $mailer, $action, $message)
+    {
+      $msg = '';
+      if ($this->parameters->getIsmail()) {
+         $msg   = $action . ' ';
+         $mailer->Subject = $msg . 'table ' . $this->tableClass->toTable() . ' (RecambioSlim)';
+         $mailer->msgHTML(date('Y-m-d H:i:s'));
+         $mailContent = '<h1>User delete api RecambioSlim</h1>';
+         $mailer->Body = $mailContent . '<p>' . json_encode($message) . '</p>';
+         if($mailer->send()) {
+             $msg .= 'Message has been sent';
+         } else {
+            $msg .= 'Mailer Error: ' . $mailer->ErrorInfo;
+         }
+      }
+      return $msg;
+    }
+
 }